+++ .docs/CHANGELOG.md
@@ -13,19 +13,23 @@ The file contains important changes from version to version.
 
 * `Assistance`
   * Add the ability to create, read, update and delete assistance
+  * Add tests for `assistance` module 
 
 * `Follower`
   * Add the ability to add, remove follower
+  * Add tests for `follower` module 
 
 * `User profile`
   * Add the ability to view and change personal info
   * Add the ability to change password
+  * Add tests for `user_profile` module 
 
 * `Authentication`
   * Add custom user model
   * Add the ability to signup, login, logout
   * Add the ability to reset password
   * Add the ability to social auth
+    * Add tests for `authentication` module 
 
 * `Initialize project`
   * Setup project and configure the structure